+++ src/MQTTClient.c
@@ -1060,7 +1060,7 @@ static int MQTTClient_connectURI(MQTTClient handle, MQTTClient_connectOptions* o
 			}
 			else
 			{
-				m->c->will->payloadlen = strlen(options->will->message);
+				m->c->will->payloadlen = (int)strlen(options->will->message);
 				source = (void*)options->will->message;
 			}
 			m->c->will->payload = malloc(m->c->will->payloadlen);
@@ -1114,7 +1114,7 @@ static int MQTTClient_connectURI(MQTTClient handle, MQTTClient_connectOptions* o
 	m->c->username = options->username;
 	m->c->password = options->password;
 	if (options->password)
-		m->c->passwordlen = strlen(options->password);
+		m->c->passwordlen = (int) strlen(options->password);
 	else if (options->struct_version >= 5 && options->binarypwd.data)
 	{
 		m->c->password = options->binarypwd.data;
